sistema de notificações bidirecionais funcionando!
- Corrigido listener de notificações no MainActivity
- Adicionado confirmação para remetente quando helper aceita
- Implementadas 40+ correções de segurança e performance
- Sistema completo de emergência funcionando

+++ app/src/main/java/com/afilaxy/domain/model/Produto.kt
@@ -1,10 +1,12 @@
 package com.afilaxy.domain.model
 
+import java.math.BigDecimal
+
 data class Produto(
     val id: String = "",
     val nome: String,
     val descricao: String,
     val imageUrl: String? = null,
-    val preco: Double? = null,
+    val preco: BigDecimal? = null,
     val categoria: String? = null
 )
